test the new avatar URL check
