+++ tests/py/test_participant.py
@@ -520,26 +520,3 @@ class Tests(Harness):
         stub = Participant.from_username(unclaimed.participant.username)
         actual = stub.resolve_stub()
         assert actual == "/on/openstreetmap/alice/"
-
-    # avatar
-
-    def test_update_avatar(self):
-        libravatar = 'https://seccdn.libravatar.org/avatar/45da67db8d78a92d35f0f5f194328b94?'
-        twitter_avatar = 'https://fake.twitter.com/path/to/avatar.jpg'
-        alice = self.make_participant('alice', email='alice@example.net')
-        alice.update_avatar()
-        assert alice.avatar_src is None
-        assert alice.avatar_url.startswith(libravatar)
-        # connect a twitter account, avatar shouldn't change
-        elsewhere = self.make_elsewhere('twitter', '1', 'alice', avatar_url=twitter_avatar)
-        alice.take_over(elsewhere)
-        assert alice.avatar_src is None
-        assert alice.avatar_url.startswith(libravatar)
-        # switch to twitter avatar
-        alice.update_avatar('twitter:')
-        assert alice.avatar_src == 'twitter:'
-        assert alice.avatar_url.startswith(twitter_avatar)
-        # check that a new call to update_avatar doesn't override avatar_src
-        alice.update_avatar()
-        assert alice.avatar_src == 'twitter:'
-        assert alice.avatar_url.startswith(twitter_avatar)
